Revert "Scripts to add data"
