+++ config_dev.py
@@ -1,0 +1,5 @@
+"""Flask configuration."""
+
+SQLALCHEMY_DATABASE_URI = "postgresql:///capstone_database"
+SQLALCHEMY_TRACK_MODIFICATIONS = False
+SECRET_KEY = "Th1s1ss3cr3t"
